updated to min SDK 21
Using onWindowFocusChanged makes the animation starts during the enter animation of the Activity. To make sure that the animation will be fully visible, you can use onEnterAnimationComplete. This method works only on API 21+ so either you set minApi to 21 or we can live with the fact that the first animation will not be visible on device running API 19. I suggest upgrading to API 21.